add GameView components
